refactor(issue-lifecycle): migrate all remaining callers to issue-lifecycle facade
Redirect imports in hitl.ts, pr-lifecycle.ts, and 16 test files from
the 7 old issue modules (issues, labels, label-lifecycle, issue-dispatch,
prd-discovery, prd-hitl, reset-labels) to the unified issue-lifecycle.ts
facade. No file outside the old modules themselves now imports directly
from them.
