+++ src/select-plan-github.test.ts
@@ -10,7 +10,7 @@ import { join } from "path";
 import { execSync } from "child_process";
 import { useTempDir } from "./test-utils.ts";
 import { selectPlanForWorktree } from "./worktree/index.ts";
-import type { PullIssueResult } from "./issues.ts";
+import type { PullIssueResult } from "./issue-lifecycle.ts";
 
 // ---------------------------------------------------------------------------
 // Helpers
